pyproject.toml
[tool.poetry]
name = "politician_collector"
version = "0.0.1"
description = ""
authors = ["sasano8 <y-sasahara@ys-method.com>"]
license = "MIT"
readme = "README.md"
# repository = "https://github.com/sasano8/asy"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
asy_rabbitmq = {git = "https://github.com/sasano8/asy_rabbitmq", rev = "main"}
uvicorn = "^0.13.4"
httpx = "^0.17.1"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.3"
elasticsearch = {extras = ["async"], version = "^7.12.0"}
pydantic = {extras = ["dotenv"], version = "^1.8.1"}
elasticsearch_dsl = "^7.3.0"
firebase_admin = "^4.5.3"
google-cloud-storage = "^1.37.1"
pandas = "^1.2.4"
xlrd = "^2.0.1"
google-cloud-firestore = "^2.1.0"
ja_text_cleaner = "^0.0.1"
sudachidict_core = "^20201223.post1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.0"
mypy = "^0.812"
pytest = "^6.2.2"
pre-commit = "^2.12.0"

# [tool.poetry.extras]
# asy = ["asy"]
# typer = ["typer"][build-system]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
